+++ flash/tabular/regression/data.py
@@ -18,7 +18,11 @@ from flash.core.data.io.input_transform import INPUT_TRANSFORM_TYPE, InputTransf
 from flash.core.utilities.imports import _PANDAS_AVAILABLE, _TABULAR_TESTING
 from flash.core.utilities.stages import RunningStage
 from flash.tabular.data import TabularData
-from flash.tabular.regression.input import TabularRegressionCSVInput, TabularRegressionDataFrameInput
+from flash.tabular.regression.input import (
+    TabularRegressionCSVInput,
+    TabularRegressionDataFrameInput,
+    TabularRegressionDictInput,
+)
 
 if _PANDAS_AVAILABLE:
     from pandas.core.frame import DataFrame
@@ -285,6 +289,116 @@ class TabularRegressionData(TabularData):
             input_cls(RunningStage.VALIDATING, val_file, **ds_kw),
             input_cls(RunningStage.TESTING, test_file, **ds_kw),
             input_cls(RunningStage.PREDICTING, predict_file, **ds_kw),
+            transform=transform,
+            transform_kwargs=transform_kwargs,
+            **data_module_kwargs,
+        )
+
+    @classmethod
+    def from_dicts(
+        cls,
+        categorical_fields: Optional[Union[str, List[str]]] = None,
+        numerical_fields: Optional[Union[str, List[str]]] = None,
+        target_field: Optional[str] = None,
+        parameters: Optional[Dict[str, Any]] = None,
+        train_dict: Optional[Dict[str, List[Any]]] = None,
+        val_dict: Optional[Dict[str, List[Any]]] = None,
+        test_dict: Optional[Dict[str, List[Any]]] = None,
+        predict_dict: Optional[Dict[str, List[Any]]] = None,
+        input_cls: Type[Input] = TabularRegressionDictInput,
+        transform: INPUT_TRANSFORM_TYPE = InputTransform,
+        transform_kwargs: Optional[Dict] = None,
+        **data_module_kwargs: Any,
+    ) -> "TabularRegressionData":
+        """Creates a :class:`~flash.tabular.regression.data.TabularRegressionData` object from the given
+        dictionary.
+
+        .. note::
+
+            The ``categorical_fields``, ``numerical_fields``, and ``target_field`` do not need to be provided if
+            ``parameters`` are passed instead. These can be obtained from the
+            :attr:`~flash.tabular.data.TabularData.parameters` attribute of the
+            :class:`~flash.tabular.data.TabularData` object that contains your training data.
+
+        The targets will be extracted from the ``target_field`` in the data frames.
+        To learn how to customize the transforms applied for each stage, read our
+        :ref:`customizing transforms guide <customizing_transforms>`.
+
+        Args:
+            categorical_fields: The fields (column names) in the dictionary containing categorical data.
+            numerical_fields: The fields (column names) in the dictionary containing numerical data.
+            target_field: The field (column name) in the dictionary containing the targets.
+            parameters: Parameters to use if ``categorical_fields``, ``numerical_fields``, and ``target_field`` are not
+                provided (e.g. when loading data for inference or validation).
+            train_dict: The dictionary to use when training.
+            val_dict: The dictionary to use when validating.
+            test_dict: The dictionary to use when testing.
+            predict_dict: The dictionary to use when predicting.
+            input_cls: The :class:`~flash.core.data.io.input.Input` type to use for loading the data.
+            transform: The :class:`~flash.core.data.io.input_transform.InputTransform` type to use.
+            transform_kwargs: Dict of keyword arguments to be provided when instantiating the transforms.
+            data_module_kwargs: Additional keyword arguments to provide to the
+              :class:`~flash.core.data.data_module.DataModule` constructor.
+
+        Returns:
+            The constructed :class:`~flash.tabular.regression.data.TabularRegressionData`.
+
+        Examples
+        ________
+
+        .. testsetup::
+
+            >>> train_data = {
+            ...     "age": [2, 4, 1],
+            ...     "animal": ["cat", "dog", "cat"],
+            ...     "weight": [6, 10, 5],
+            ... }
+            >>> predict_data = {
+            ...     "animal": ["dog", "dog", "cat"],
+            ...     "weight": [7, 12, 5],
+            ... }
+
+        We have dictionaries ``train_data`` and ``predict_data``.
+
+        .. doctest::
+
+            >>> from flash import Trainer
+            >>> from flash.tabular import TabularRegressor, TabularRegressionData
+            >>> datamodule = TabularRegressionData.from_dicts(
+            ...     "animal",
+            ...     "weight",
+            ...     "age",
+            ...     train_dict=train_data,
+            ...     predict_dict=predict_data,
+            ...     batch_size=4,
+            ... )
+            >>> model = TabularRegressor.from_data(datamodule, backbone="tabnet")
+            >>> trainer = Trainer(fast_dev_run=True)
+            >>> trainer.fit(model, datamodule=datamodule)  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
+            Training...
+            >>> trainer.predict(model, datamodule=datamodule)  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
+            Predicting...
+
+        .. testcleanup::
+
+            >>> del train_data
+            >>> del predict_data
+        """
+        ds_kw = dict(
+            categorical_fields=categorical_fields,
+            numerical_fields=numerical_fields,
+            target_field=target_field,
+            parameters=parameters,
+        )
+
+        train_input = input_cls(RunningStage.TRAINING, train_dict, **ds_kw)
+        ds_kw["parameters"] = train_input.parameters if train_input else parameters
+
+        return cls(
+            train_input,
+            input_cls(RunningStage.VALIDATING, val_dict, **ds_kw),
+            input_cls(RunningStage.TESTING, test_dict, **ds_kw),
+            input_cls(RunningStage.PREDICTING, predict_dict, **ds_kw),
             transform=transform,
             transform_kwargs=transform_kwargs,
             **data_module_kwargs,